fastapi/requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
PyMySQL==1.1.1
peewee==3.17.6
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
itsdangerous==2.1.2
passlib==1.7.4
python-multipart==0.0.12
cryptography==43.0.1
pycryptodomex==3.20.0
pyjwt==2.9.0
strenum==0.4.15
pdfplumber==0.10.4
pillow==10.3.0
aspose-slides==24.9.0; platform_machine == 'x86_64'
minio==7.2.4
redis==5.0.3
openpyxl==3.1.2
pandas==2.2.2
elasticsearch==8.13.0
elasticsearch_dsl==8.13.0
elastic-transport==8.13.0
pyvi==0.1.1
tiktoken==0.8.0
pypdf==5.0.1
datrie==0.8.2
openai==1.54.4
PyMuPDF==1.24.9
huggingface_hub==0.26.2
onnxruntime==1.20.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
sentence_transformers==3.4.1
umap_learn==0.5.6
httpx==0.27.2
langchain==0.3.13
langgraph==0.2.60
langchain_openai==0.2.8
langchain_core==0.3.28
scholarly==1.7.11
duckduckgo_search==7.1.1
streamlit==1.41.1
markdown_to_json==2.1.2
graspologic==3.4.1
future==1.0.0
markdown_to_json==2.1.2
editdistance==0.8.1
tavily-python
requests
bs4